UI and Loading Scene Ready

+++ Assets/Scripts/Model/Interfaces/IBubblesObserver.cs
@@ -5,7 +5,7 @@ namespace BubblesShoot.Model.Interfaces
 {
     public interface IBubblesObserver
     {
-        void UpdateView(List<List<BubbleCell>> bubbles, float score);
+        void UpdateView(List<List<BubbleCell>> bubbles, int score);
     }
 
     public interface IBubblesObservable
